fix(behavior_path_planner): move module output interface definition
